Change font engine to SFE++
Changed the font engine to Stamp Font Engine ++ font engine created by @-REX-. I switched to this one because it's faster and overall better. I also changed the font to work sans.

The conversion isn't complete, I still have to fix all the text alignment, and the auto width for the nodes.